Added HTML versions for use in Inkdit
The markdown versions copy and paste nicely into Inkdit but you can use
these HTML versions to save a step. Note: The signing lines in Inkdit
include the name, Company (where appropriate), timestamp. At the time
of this commit, we don't automatically include title when signing on
behalf of an organization but I expect that feature will be in
production before anyone uses these contracts.